refactor: extract session broadcast from AgentWorkspace
