+++ src/components/agent/use-session-list-broadcast-types.ts
@@ -1,0 +1,9 @@
+import type { HermesSessionInfo } from "../../lib/tauri";
+
+export type UseSessionListBroadcastDependencies = {
+  hermesSessionItems: HermesSessionInfo[];
+  hermesSessionsHydrated: boolean;
+  selectedHermesSessionId: string | undefined;
+  waitingSessionIds: Set<string>;
+  workingSessionIds: Set<string>;
+};
